requirements.txt
opencv-python==4.8.0.76
numpy==1.24.3
tensorflow==2.13.0
keras==2.13.1
mediapipe==0.10.3
pillow==10.0.0 